requirements.txt
paho-mqtt
fastapi
uvicorn
pytest-html
pylint